+++ src/main/java/com/snackecommerce/product/dto/CouponResponse.java
@@ -6,6 +6,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import com.snackecommerce.product.entity.Coupon.CouponType;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -16,8 +17,8 @@ public class CouponResponse {
     private Long id;
     private String code;
     private CouponType type;
-    private Double discountValue;
-    private Double minOrderAmount;
+    private BigDecimal discountValue;
+    private BigDecimal minOrderAmount;
     private Boolean active;
     private LocalDateTime validFrom;
     private LocalDateTime validTill;
